% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{add_edges}
\alias{add_edges}
\title{Edges insertion}
\usage{
add_edges(.graph, listOfEdges)
}
\arguments{
\item{.graph}{an ArangoGraph handler}

\item{listOfEdges}{a list containing the edges to be added in the graph}
}
\description{
Adds a set of edges to the given graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
