% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{add_to_graph}
\alias{add_to_graph}
\title{Adds a set of verticies to the given graph}
\usage{
add_to_graph(.graph, listOfEdges)
}
\arguments{
\item{listOfEdges}{a list containing the edges to be added in the graph}
}
\description{
Adds a set of verticies to the given graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
