% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph_concrete.R
\name{all_graph}
\alias{all_graph}
\title{Get complete graph}
\usage{
all_graph(.graph)
}
\description{
Returns the graph vertices and edges for the given ArangoGraph (BE CAREFUL, the grap may
 contains lot of elements)
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
