% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_database.R
\name{arango_database}
\alias{arango_database}
\title{Get or create database}
\usage{
arango_database(.connection, name = "_system", createOnFail = FALSE)
}
\arguments{
\item{.connection}{the ArangoConnection handler}

\item{name}{the name of the database, the default is "_system" that is ArangoDB default one}

\item{createOnFail}{if the database were not found creates it. If the parameter is set to TRUE
and the default database is requested the function fails. (default FALSE)}
}
\description{
Return an object representing the database with the given name:
the object must be used to handle requests to the database.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
