% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{arango_graph}
\alias{arango_graph}
\title{Returns the graph identified with the given name}
\usage{
arango_graph(.database, name, createOnFail = FALSE)
}
\arguments{
\item{name}{the name of the graph}
}
\description{
Returns the graph identified with the given name
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
