% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_collection.R
\name{collection}
\alias{collection}
\title{Get or create a collection}
\usage{
collection(.database, name, createOnFail = FALSE)
}
\arguments{
\item{name}{the name of the collection}

\item{createOnFail}{if the collection were not found creates it (default FALSE)}

\item{.element}{a valid ArangoDatabase object}
}
\value{
an object representing the collection with the given name
}
\description{
Return an object representing the collection with the given name into the specified the database:
the object must be used to handle requests to the collection.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
