% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{collection_update}
\alias{collection_update}
\title{Execute document updates}
\usage{
collection_update(.data)
}
\arguments{
\item{the}{ArangoDocument to be updated on the server}
}
\value{
the ArangoDocument updated but not yet consistent with the server image
}
\description{
Excecutes the update of a document on the server
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
