% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_collection.R
\name{collections}
\alias{collections}
\title{Returns all the collections belonging to the giving database}
\usage{
collections(.database, includeSystem = FALSE)
}
\description{
Returns all the collections belonging to the giving database
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
