% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_connection.R
\name{connect}
\alias{connect}
\title{Creates a connection object to be used for subsequent requests to the given server.}
\usage{
connect(host, port)
}
\arguments{
\item{host}{the server address where the ArangoDB server is up and running}

\item{port}{the server port where the ArangoDB server is up and running}
}
\description{
Creates a connection object to be used for subsequent requests to the given server.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
