% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{connections}
\alias{connections}
\title{Get complete graph}
\usage{
connections(.graph)
}
\description{
Returns the graph vertices and edges for the given ArangoGraph (BE CAREFUL, the grap may
 contains lot of elements)
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
