% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_database.R
\name{database}
\alias{database}
\title{Return a wrapper to the given database, if any for the given connection}
\usage{
database(.connection, name = "_system", createOnFail = FALSE)
}
\arguments{
\item{.connection}{the server connection to the ArangoDB instance}

\item{name}{the name of the database to which connect to, the default is "_system" that is
the default of ArangoDB}

\item{createOnFail}{if the database were not found creates it. If the parameter is set to TRUE
and the default database is requested the function fails. Default is FALSE}
}
\description{
Return a wrapper to the given database, if any for the given connection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
