% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_database.R
\name{databases}
\alias{databases}
\title{Get all databases}
\usage{
databases(.connection, includeSystem = FALSE)
}
\arguments{
\item{.connection}{the ArangoConnection handler}

\item{includeSystem}{TRUE iff the system databases must be included in the results, FALSE
otherwise}
}
\value{
a character vector with all the databases available in the server
}
\description{
Returns all the databases available in the server identified by the current connection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
