% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{delete}
\alias{delete}
\title{Delete document}
\usage{
delete(.document)
}
\arguments{
\item{.document}{the ArangoDocument to be removed from the collection}
}
\value{
TRUE iff the document has been removed from the collection
}
\description{
Deletes the given document
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
