% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{document_insert}
\alias{document_insert}
\title{Create new documents}
\usage{
document_insert(.collection, key)
}
\arguments{
\item{.collection}{the ArangoCollection that handles the collection to be updated}

\item{key}{the key that will be used as unique identifier within the collection for the new document}
}
\value{
an ArangoDocument representing the new document data
}
\description{
Creates a new document of this collection and returns it
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
