% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_connection.R
\docType{data}
\name{.aRango_connection}
\alias{.aRango_connection}
\title{An ArangoConnection is a class that contains and manages the connection with one specific 
instance of ArangoDB. Basically this object must be used to get databases collections, graphs
or to interact in other ways with an existing instance of the db.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_connection
}
\arguments{
\item{host}{the address on which the Arango instance is running}

\item{port}{the port on the server on which the Arango instance is running}

\item{database}{OPTIONAL, the name of the database to which connect

Returns the server indicated by the instance running on this connection


Returns the version of the server indicated by the instance running on this connection


Returns the license indicated by the instance running on this connection


Returns a string in the form of "http://<localhost>:<port>" useful to execute new requests}
}
\description{
An ArangoConnection is a class that contains and manages the connection with one specific 
instance of ArangoDB. Basically this object must be used to get databases collections, graphs
or to interact in other ways with an existing instance of the db.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
Creates a new connection to a server running ArangoDB
}
\keyword{datasets}
