% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_database.R
\docType{data}
\name{.aRango_database}
\alias{.aRango_database}
\title{An ArangoDatabase is a class that wraps all the operations and interaction methods with
a database existing within the specific ArangoConnection.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_database
}
\arguments{
\item{connection}{an ArangoConnection with an ArangoDB server}

\item{dbname}{the database name to which connect to 

Returns the name of the db wrapped by this object}
}
\description{
An ArangoDatabase is a class that wraps all the operations and interaction methods with
a database existing within the specific ArangoConnection.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
Creates a new interface to some database available through the given connection

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns TRUE iff this object is connected to a system database, FALSE otherwise

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the identifier of the database within the connected server

Gabriele Galatolo, g.galatolo(at)kode.srl
}
\keyword{datasets}
