% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\docType{data}
\name{.aRango_graph}
\alias{.aRango_graph}
\title{The ArangoGraph class encapsulate the representation and the methods useful to manage
the retrieval, representation and updates of specific node and vertices belonging to Arango graph.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_graph
}
\description{
Creates a new graph from a one existing on the database
}
\details{
Returns the name of the graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the identifier of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the current revision of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the edges definition of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
}
\keyword{datasets}
