% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\docType{data}
\name{.aRango_subgraph}
\alias{.aRango_subgraph}
\title{The ArangoGraph class encapsulate the representation and the methods useful to manage
the retrieval, representation and updates of specific node and vertices belonging to Arango graph.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_subgraph
}
\description{
Returns a list of matrices that contains, for each relation between the nodes, an adjacency matrix
for that relation.
For example, given "a" -friend_of-> "b", "c" -spouse-> "b" the results will be:
}
\details{
a   b   c
            a  0   1   0
friend_of = b  0   0   0
            c  0   0   0
            
               a   b   c
            a  0   0   0
spouse    = b  0   0   0
            c  0   1   0

If "complete" option is set to TRUE the content of the matrix will be the edge key within the ArangoDB
server: looking to the previous example

a         b         c
            a  0   friend_of/112   0
friend_of = b  0         0         0
            c  0         0         0
            
               a        b       c
            a  0        0       0
spouse    = b  0        0       0
            c  0   spouse/231   0
}
\keyword{datasets}
