% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_functions.R
\name{.drop_collection}
\alias{.drop_collection}
\title{Internal function for collection removal}
\usage{
.drop_collection(.element)
}
\description{
Internal function for collection removal
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
