% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_functions.R
\name{.drop_database}
\alias{.drop_database}
\title{Internal function for database removal}
\usage{
.drop_database(.element)
}
\description{
Internal function for database removal
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
