% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_functions.R
\name{.drop_graph}
\alias{.drop_graph}
\title{Internal function for graph removal}
\usage{
.drop_graph(.element)
}
\description{
Internal function for graph removal
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
