% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{.element_coloring}
\alias{.element_coloring}
\title{Default elements coloring}
\usage{
.element_coloring(edges, palette = RColorBrewer::brewer.pal(8,
  "Pastel1"))
}
\description{
Calculates different coloring for the elements of the graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
