% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_functions.R
\name{drop}
\alias{drop}
\title{Drop an existing database or collection}
\usage{
drop(.element)
}
\arguments{
\item{.element}{can be an ArangoDatabase OR ArangoCollection, other objects raise exception}
}
\description{
Drop an existing database or collection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
