% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{edge}
\alias{edge}
\title{Creates a list with the _from/_to indication used to define an anonymous edge, i.e. still without
relation for which the edge is defined. If indicated, the attributes (in key-value form) for the resulting anonymous edge}
\usage{
edge(definition, ...)
}
\arguments{
\item{definition}{a list containing a `_from` and a `_to` key}

\item{...}{pairs of key-values representing attributes of this edge}
}
\description{
Creates a list with the _from/_to indication used to define an anonymous edge, i.e. still without
relation for which the edge is defined. If indicated, the attributes (in key-value form) for the resulting anonymous edge
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
