% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{edge_definition}
\alias{edge_definition}
\title{Adds an edge definition to this graph}
\usage{
edge_definition(.graph, fromCollection, relation, toCollection)
}
\description{
Adds an edge definition to this graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
