% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_enumeration.R
\name{enum}
\alias{enum}
\title{Function that allows to define an enumeration in Java-like style.}
\usage{
enum(...)
}
\arguments{
\item{...}{unique values that define the enumeration set}
}
\description{
Function that allows to define an enumeration in Java-like style.
}
\references{
https://stackoverflow.com/questions/33838392/enum-like-arguments-in-r
}
\author{
AEF (StackOverflow user)
}
