% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{find_edge}
\alias{find_edge}
\title{Returns the edge in this collection connecting the document passed as argument, if any.}
\usage{
find_edge(.collection, from, to)
}
\description{
Returns the edge in this collection connecting the document passed as argument, if any.
}
