% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{\%->\%}
\alias{\%->\%}
\title{Creates a list representing an empty edge with an outbound from v1 to v2}
\usage{
v1 \%->\% v2
}
\arguments{
\item{v1}{source document of the edge}

\item{v2}{destination document of the edge}
}
\description{
Creates a list representing an empty edge with an outbound from v1 to v2
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
