% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{\%gt\%}
\alias{\%gt\%}
\title{Creates a string expression representing an AQL filter "greater than" clause for
numeric values}
\usage{
expr \%gt\% value
}
\arguments{
\item{expr}{the expression to be left as given in the resulting string}

\item{value}{the expression to be evaluated and printed in the resulting string}
}
\description{
Creates a string expression representing an AQL filter "greater than" clause for
numeric values
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
