% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{\%owns\%}
\alias{\%owns\%}
\title{Assign a relation, i.e. an edge collection, to an anonymous edge, i.e. a list containing at least _from/_to
attributes.}
\usage{
relation \%owns\% edgeDefinition
}
\arguments{
\item{relation}{a string that represents an edge collection}

\item{edges}{a list containing an anonymous edge (TODO, this must be a list of anonymous edges)}
}
\description{
Assign a relation, i.e. an edge collection, to an anonymous edge, i.e. a list containing at least _from/_to
attributes.
}
\examples{
"requires" \%:\% edge(data_science \%->\% math) returns a list(collection=..., edge=...) where 
edge contains a _from and a _to valid verticies

}
\seealso{
edge
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
