% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{graphs}
\alias{graphs}
\title{Get all graphs}
\usage{
graphs(.database)
}
\arguments{
\item{.database}{the AranagoDatabase handler}
}
\description{
Returns all the graphs belonging to the giving database
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
