% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{insert}
\alias{insert}
\title{Creates a new document of this collection and returns it}
\usage{
insert(.collection, key)
}
\arguments{
\item{.collection}{the collection where add the new document}

\item{key}{the key that will be used as unique identifier within the collection for the new document}
}
\description{
Creates a new document of this collection and returns it
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
