% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{remove_from_graph}
\alias{remove_from_graph}
\title{Edges removal}
\usage{
remove_from_graph(.graph, listOfEdges)
}
\arguments{
\item{.graph}{the graph affected by the deletion}

\item{listOfEdges}{a list of lists containing edges information}
}
\value{
the ArangoGraph object of the structure affected by the change
}
\description{
Removes a collection of elements from the given graph.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
