% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph_concrete.R
\name{traversal}
\alias{traversal}
\title{Traversal of a graph}
\usage{
traversal(.graph, vertices, depth = 1, edges = NULL)
}
\arguments{
\item{vertices}{}
}
\description{
Execute the traversal of a graph OR starting from a given set of edges
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
