% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{visualize}
\alias{visualize}
\title{Graph visualization}
\usage{
visualize(.element)
}
\arguments{
\item{.element}{can be an 'ArangoGraphConcrete' or a list containing other visualzation
options. This list is automatically created when this method is inside a pre-visualization
pipe}
}
\value{
the visualization of the graph
}
\description{
Returns the visualization of the given graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
