% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{viz_edges}
\alias{viz_edges}
\title{Set visualization edges options}
\usage{
viz_edges(.element, directions = NULL)
}
\arguments{
\item{.element}{can be an 'ArangoGraphConcrete' or a list containing other visualzation
options. This list is automatically created when this method is inside a pre-visualization
pipe}

\item{directions}{a list containing the direction to be visualized on the edge for each type
of relation contained into the graph, e.g. in a graph containing "friend_of" and "spouse" this parameter
can be list(friend_of="to", spouse="from")}
}
\value{
a list containing the visualization options updated
}
\description{
Allows the user to set some options regarding the visualization of the edges of the graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
