% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{viz_legend}
\alias{viz_legend}
\title{Set graph legend}
\usage{
viz_legend(.element, nodes = TRUE, edges = FALSE)
}
\arguments{
\item{.element}{can be an 'ArangoGraphConcrete' or a list containing other visualzation
options. This list is automatically created when this method is inside a pre-visualization
pipe}

\item{nodes}{a boolean that indicate whether show or not the legend for nodes}

\item{edges}{a boolean that indicate whether show or not the legend for edges}
}
\value{
a list containing the visualization options updated
}
\description{
Allows the user to the legend for this graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
