% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{viz_vertices}
\alias{viz_vertices}
\title{Set visualization nodes options}
\usage{
viz_vertices(.element, icons = NULL, colors = NULL)
}
\arguments{
\item{.element}{can be an 'ArangoGraphConcrete' or a list containing other visualzation
options. This list is automatically created when this method is inside a pre-visualization
pipe}

\item{icons}{a list containing the icon name (from FontAwesome) to be set for each node type}

\item{colors}{a list containing the colors to assign to each node type}
}
\value{
a list containing the visualization options updated
}
\description{
Allows the user to set some options regarding the visualization of the nodes of the graph
}
\examples{
graph \%>\% viz_vertices(icons=list(employee="f007", company="f275"), colors=list(employee="red", company="blues"))

}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
