% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_visualization.R
\name{viz_width}
\alias{viz_width}
\title{Set graph width}
\usage{
viz_width(.element, width)
}
\arguments{
\item{.element}{can be an 'ArangoGraphConcrete' or a list containing other visualzation
options. This list is automatically created when this method is inside a pre-visualization
pipe}

\item{width}{the width of the resulting graph in CSS style, e.g. "90\%"}
}
\value{
a list containing the visualization options updated
}
\description{
Allows the user to set the width for the canvas of the graph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
