% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_aql.R
\name{aql}
\alias{aql}
\title{AQL query wrapper}
\usage{
aql(.element, statement)
}
\arguments{
\item{.element}{a valid ArangoDatabase object}

\item{statement}{an AQL statement}
}
\description{
Create an R function that interprets the statement as AQL query. Every
bindvar of the statement will become a mandatory parameter of the function.
}
\examples{
paperWithTitle <- db \%>\% aql("FOR p IN Papers FILTER p.title==@search RETURN p")
paper <- paperWithTitle(search="A study in scarlet")

}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
