% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_collection.R
\name{collection}
\alias{collection}
\title{Return a wrapper to the given database, if any for the given connection}
\usage{
collection(.database, name, createOnFail = FALSE, createOption = NULL)
}
\arguments{
\item{name}{the name of the collection to which connect to}

\item{createOnFail}{if the collection were not found creates it. Default is FALSE}

\item{.connection}{the server connection to the ArangoDB instance}

\item{createOptions}{if a list is provided tells which options must be used when creating the new collection.
Valid options are 'waitForSync' (BOOLEAN), 'isSystem' (BOOLEAN), 'type' (collection_type).}
}
\description{
Return a wrapper to the given database, if any for the given connection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
