% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_collection.R
\name{collections}
\alias{collections}
\title{Get all collections name}
\usage{
collections(.database, includeSystem = FALSE)
}
\arguments{
\item{includeSystem}{if TRUE include the system databases (default FALSE)}

\item{.element}{a valid ArangoDatabase object}
}
\value{
a character vector with all the collections contained in the given database
}
\description{
Returns all the collections contained in the given database
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
