% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{connections}
\alias{connections}
\title{Returns the entire graph passed as first argument}
\usage{
connections(.graph)
}
\description{
Returns the entire graph passed as first argument
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
