% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\name{connections}
\alias{connections}
\title{Returns the graph vertices and edges for the given ArangoGraph}
\usage{
connections(.graph)
}
\description{
Returns the graph vertices and edges for the given ArangoGraph
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
