% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_database.R
\name{databases}
\alias{databases}
\title{Returns all the databases available on the server identified by the current connection}
\usage{
databases(.connection, includeSystem = FALSE)
}
\arguments{
\item{.connection}{the server connection to the ArangoDB instance}

\item{includeSystem}{TRUE iff the system databases must be included in the results, FALSE
otherwise}
}
\description{
Returns all the databases available on the server identified by the current connection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
