% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{documents}
\alias{documents}
\title{Returns all the documents belonging to the giving collections}
\usage{
documents(.collection)
}
\description{
Returns all the documents belonging to the giving collections
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
