% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\name{documents}
\alias{documents}
\title{Get all documents}
\usage{
documents(.collection)
}
\value{
all the documents belonging to the given collection
}
\description{
Returns all the documents belonging to the given collection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
