% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_documents.R
\docType{data}
\name{.aRango_document}
\alias{.aRango_document}
\title{The ArangoDocument class encapsulate the representation and the methods useful to manage
the retrieval, representation and updates of specific documents belonging to Arango collections.
Each document is strictly related to some collection and within the environment SHOULD exists
only one copy of this object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_document
}
\description{
Initialize the arango document
}
\keyword{datasets}
