% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_graph.R
\docType{data}
\name{.aRango_graph}
\alias{.aRango_graph}
\title{An ArangoGraph is a class where instances are used to handle the interaction with
real graphs on the server.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
.aRango_graph
}
\description{
An ArangoGraph is a class where instances are used to handle the interaction with
real graphs on the server.
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
Creates a new graph from a one existing on the database

Returns the name of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the identifier of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the current revision of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the edges definition of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
Returns the orphan collections of the graph

Gabriele Galatolo, g.galatolo(at)kode.srl
}
\keyword{datasets}
