% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_functions.R
\name{drop}
\alias{drop}
\title{Drop an handler}
\usage{
drop(.element)
}
\arguments{
\item{.element}{can be an ArangoDatabase or an ArangoCollection or an ArangoGraph}
}
\value{
TRUE iff the element has been deleted, FALSE otherwise
}
\description{
Drop an existing database, graph or collection
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
