% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{\%lt\%}
\alias{\%lt\%}
\title{Creates a string expression representing a filter clause for AQL}
\usage{
expr \%lt\% value
}
\arguments{
\item{expr}{the expression to left as given in the resulting string}

\item{value}{the expression to be evaluated and printed in the resulting string}
}
\description{
Creates a string expression representing a filter clause for AQL
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
