% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arango_common_operators.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Creates a list representing an empty edge with an outbound from v2 to v1}
\usage{
v1 \%<-\% v2
}
\arguments{
\item{v1}{destination document of the edge}

\item{v2}{source document of the edge}
}
\description{
Creates a list representing an empty edge with an outbound from v2 to v1
}
\author{
Gabriele Galatolo, g.galatolo(at)kode.srl
}
